pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "mayconcorreatinti",email = "maycon326754@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.127.0,<0.128.0)",
    "mysql-connector-python (>=9.5.0,<10.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "ruff (>=0.14.13,<0.15.0)",
    "pymongo (>=4.16.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 78

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true